+++ utils/log.py
@@ -1,53 +1,82 @@
 import logging
+import os
+import sys
+from functools import partial, partialmethod
+from inspect import stack
 
-import coloredlogs
-import verboselogs
+import yaml
+from bunch import Bunch
+from emoji import emojize
+from loguru import logger
+from rich.console import Console
+from yaml import FullLoader as Full_Loader
 
-from utils.exception import to_str
+from about import name as about_name
 
-# FIXME DEBUG level not working
+emoji = partial(emojize, use_aliases=True)
+console = Console()
+
+
+class Formatter:
+    @staticmethod
+    def info(name):
+        for s in stack():
+            if name in s.filename.replace('_', '-'):
+                return s
+        return None
+
+    @classmethod
+    def apply(cls, record):
+        # s = cls.info(record['name'])
+        # record['called'] = Bunch(filename=path.basename(s.filename), function=s.function, lineno=s.lineno, icon=emoji(':computer:'))
+        record['called'] = Bunch(filename=record['file'].name, function=record['function'], lineno=record['line'], icon=emoji(':computer:'))
+        record['elapsed'] = Bunch(time=record['elapsed'], icon=emoji(':alarm_clock:'))
+        record['message'] = emoji(record['message'])
 
 
 class Log:
-    """Wrapper class for coloured and verbuse logs."""
-
-    levels = {}
-    default = None
+    default_icon = ':small_orange_diamond:'
 
     @classmethod
-    def init(cls, default, levels):
+    def init(cls, config):
         """Set the default and levels and initialize the log manager.
 
         :param cls: Log class.
-        :param default: default log level
-        :param levels: log levels
+        :param config: Path of the config filename
+        :param clear: Clear the previous logs.
         """
-        verboselogs.install()
-        coloredlogs.install()
-        cls.default = default
-        cls.levels = {module: level for module, level in levels}
-        logging.get_logger = logging.getLogger
-        for module, level in levels:
-            logging.get_logger(module).setLevel(level)
+        with open(config) as cfg_file:
+            cfg = yaml.load(cfg_file, Loader=Full_Loader)
+        levels = cfg.get('levels', [])
+        for level in levels:
+            level['icon'] = emoji(level.get('icon', cls.default_icon))
+            name = level.get('name', None)
+            if name:
+                klass = type(logger)
+                setattr(klass, name.lower(), partialmethod(klass.log, name))
+        hdls = []
+        for sink_data in cfg.get('sinks', []):
+            if sink_data.get('enabled', True):
+                klass = sink_data.get('klass', None)
+                if klass:
+                    if klass.lower() == 'stdout':
+                        sink = sys.stdout
+                    elif klass.lower() == 'stderr':
+                        sink = sys.stderr
+                    else:
+                        sink = klass.format(name=name)
+                        if sink_data.get('clear', True) and os.path.exists(sink):
+                            os.remove(sink)
+                    h = dict(sink=sink, **sink_data.get('args', {}))
+                    hdls.append(h)
+        logger.configure(handlers=hdls, levels=cfg.get('levels', {}), patcher=Formatter.apply)
+
+        for level in levels:
+            cls.get('log').info(f"Found additional log level customization: {level['icon']:<3} {level['name']}")
 
     @classmethod
-    def get(cls, name):
-        """Return the initialized logger with the module name.
-
-        :param cls: Log class.
-        :param name: module name
-        :returns: logger instance
-        """
-        level = cls.levels.get(name, cls.default)
-        logger = logging.get_logger(name)
-
-        def __exception(message, exception):
-            logger.error(message)
-            logger.debug(to_str(exception))
-        logger.exception = __exception
-
-        logger.setLevel(level)
-        return logger
+    def get(cls, name=about_name):
+        return logger.bind(context=name).opt(colors=True)
 
     @staticmethod
     def get_levels():
@@ -55,4 +84,4 @@ class Log:
 
         :returns: list of string
         """
-        return logging._levelToName.values()
+        return list(logging._levelToName.values())