+++ CHANGELOG.md
@@ -24,3 +24,4 @@
 
 - Added option to force the use of HTTPS instead of HTTP.
 - Add JWT Authentication.
+- Enhanced Log configuration.